+++ app/src/main/java/com/lilun/passionlife/cloudplatform/ui/fragment/AddOrganizationFragment.java
@@ -1,6 +1,5 @@
 package com.lilun.passionlife.cloudplatform.ui.fragment;
 
-import android.content.Context;
 import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -15,23 +14,26 @@ import com.lilun.passionlife.cloudplatform.bean.Event;
 import com.lilun.passionlife.cloudplatform.bean.IsInherited;
 import com.lilun.passionlife.cloudplatform.bean.Organization;
 import com.lilun.passionlife.cloudplatform.common.Constants;
+import com.lilun.passionlife.cloudplatform.common.PicloadManager;
 import com.lilun.passionlife.cloudplatform.custom_view.CircleImageView;
 import com.lilun.passionlife.cloudplatform.custom_view.ExtendItem;
 import com.lilun.passionlife.cloudplatform.custom_view.RegItemView;
 import com.lilun.passionlife.cloudplatform.net.retrofit.ApiFactory;
 import com.lilun.passionlife.cloudplatform.net.rxjava.PgSubscriber;
+import com.lilun.passionlife.cloudplatform.ui.App;
+import com.lilun.passionlife.cloudplatform.utils.StringUtils;
 import com.lilun.passionlife.cloudplatform.utils.ToastHelper;
 import com.orhanobut.logger.Logger;
 
 import org.greenrobot.eventbus.EventBus;
 import org.greenrobot.eventbus.Subscribe;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
 import butterknife.Bind;
 import butterknife.OnClick;
+import okhttp3.RequestBody;
 
 /**
  * Created by youke on 2016/6/22.
@@ -58,7 +60,7 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
     @Bind(R.id.input_orgi_desc)
     RegItemView inputOrgiDesc;
 
-    private List<Organization> departments;
+    private List<Organization> departments = new ArrayList<>();
     private String orgaName;
     private Organization orgna;
     private String orgaDesc;
@@ -66,6 +68,7 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
     private int size;
     private Organization newDepartment;
     private View view;
+    private Bitmap icon;
 
     @Override
     public View setView() {
@@ -75,26 +78,16 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
     }
 
 
-    @Override
-    public void onAttach(Context context) {
-        super.onAttach(context);
-        departments = new ArrayList<>();
-    }
 
     @Override
     public void onStart() {
         super.onStart();
+        rootActivity.setTitle(App.app.getString(R.string.add_orgi));
         exvAddOrgi.setOnClickListen(this);
     }
 
 
-    /**
-     * 更新组织，需要设置初始化数据
-     */
-    private void setInitData() {
-        inputOrgiName.setInput(TextUtils.isEmpty(orgna.getName()) ? "" : orgna.getName());
-        inputOrgiDesc.setInput(TextUtils.isEmpty(orgna.getDescription()) ? "" : orgna.getDescription());
-    }
+
 
 
     /**
@@ -123,7 +116,10 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
     public void choisHead(Event.choiseHeadPic event){
         Bitmap bitmap = event.getBitmap();
         if (bitmap!=null){
-            ivHead.setImageBitmap(bitmap);
+            StringUtils.getBytesFromBitmap(bitmap);
+//            icon = bitmap;
+//            ivHead.setImageBitmap(bitmap);
+
         }
     }
 
@@ -134,36 +130,39 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
      * @param id
      */
     private void saveDepartment(String id) {
-        size = departments.size();
+        if (departments==null && departments.size()==0){return;}
+        List<Organization>  depts = new ArrayList<>();
         for (int i = 0; i < departments.size(); i++) {
             Organization depart = departments.get(i);
             depart.setId(id + Constants.special_orgi_department + "/" + depart.getName());
             depart.setParentId(id + Constants.special_orgi_department);
             Logger.d("department id  = " + id + Constants.special_orgi_department + "/" + depart.getName());
             Logger.d("department parentid  = " + id + "/" + depart.getName() + Constants.special_orgi_department);
-            rootActivity.addSubscription(ApiFactory.postOrganization(departments.get(i)), new PgSubscriber<Organization>(rootActivity) {
-                @Override
-                public void on_Next(Organization organizationBean) {
-                    then();
-
-                }
-
-
-            });
+            depts.add(depart);
         }
+
+
+        rootActivity.addSubscription(ApiFactory.postOrganizations(depts), new PgSubscriber<List<Organization>>(rootActivity) {
+            @Override
+            public void on_Next(List<Organization> organizationBean) {
+                ToastHelper.get().showShort(App.app.getString(R.string.add_department_success));
+                rootActivity.backStack();
+                postDeptIcon();
+            }
+        });
     }
 
-    private void then() {
-        if (dex == size - 1) {
-            rootActivity.backStack();
-            return;
-        }
-        dex++;
+    /**
+    *新增部门的图标
+    */
+    private void postDeptIcon() {
+        //TODO 上传部门的图标
+
     }
 
 
     /**
-     * 存 新增的组织
+     * 存新增的组织
      */
     private void saveOrgi() {
         orgna = new Organization();
@@ -171,26 +170,49 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
         orgna.setId(orgiId + "/" + orgaName);
         orgna.setName(orgaName);
         orgna.setParentId(orgiId);
-        Logger.d("orgi name = " + orgaName);
-        Logger.d("orgi desc = " + orgaDesc);
+        Logger.d("isInherited" + orgna.isInherited());
         orgna.setDescription(orgaDesc);
         //TODO 设置icon
         rootActivity.addSubscription(ApiFactory.postOrganization(orgna), new PgSubscriber<Organization>(rootActivity) {
             @Override
             public void on_Next(Organization orga) {
-
-                if (departments != null && departments.size() != 0) {
-                    String id = orga.getId() + Constants.special_orgi_department;
-                    rootActivity.setIsInherited(id, new IsInherited(orga.isInherited()));
-                    Logger.d(" set isInherited id = "+id+"---"+"isInherited = "+orga.isInherited());
-                    saveDepartment(orga.getId());
-                } else {
-                    rootActivity.backStack();
+                ToastHelper.get().showShort(orgna.getName()+"组织新增成功");
+                String id = orga.getId()+Constants.special_orgi_department;
+                rootActivity.setIsInherited(id,new IsInherited(orga.isInherited()));
+                if (icon!=null){
+                    saveIcon(orga);
                 }
+
+                //不是集成就添加自己的部门
+                if (!orga.isInherited()){
+                    saveDepartment(orga.getId());
+                }
+
             }
 
-
+            @Override
+            public void on_Error() {
+                super.on_Error();
+                ToastHelper.get().showShort(orgna.getName()+"组织新增失败");
+            }
         });
+    }
+
+    /**
+    *新增组织Icon
+     */
+    private void saveIcon(Organization orga) {
+        if (icon!=null){
+            String orgaId = orga.getId();
+            RequestBody requestBody = PicloadManager.getUploadIconRequestBody(icon);
+            rootActivity.addSubscription(ApiFactory.postOrganizationIcon(orgaId, requestBody), new PgSubscriber(rootActivity) {
+                @Override
+                public void on_Next(Object o) {
+                    ToastHelper.get().showShort("组织图标上传成功");
+                }
+            });
+        }
+
     }
 
 
@@ -239,9 +261,7 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
      *新增what
      */
     public void onAddWhatClick() {
-//        saveStateToArguments();
         Event.OpenNewFragmentEvent event = new Event.OpenNewFragmentEvent(new AddDepartmentFragment(), mCx.getString(R.string.add_department));
-//        event.setBundle(savedState);
         EventBus.getDefault().post(event);
 
     }
@@ -252,11 +272,9 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
      */
     @Subscribe
     public void addDepartment(Event.addNewDepartment event) {
-        Logger.d("   get  a new depart");
         newDepartment = event.getDepartment();
-//        departments.add(department);
-//        Logger.d(" lis size = " +departments.size());
-//        exvAddOrgi.setListviewData(new ExtDeptAdapter(departments, mCx));
+        departments.add(newDepartment);
+        Logger.d(" lis size = " +departments.size());
     }
 
 
@@ -267,7 +285,7 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
         super.onSaveState(outState);
         outState.putString("organiName", inputOrgiName.getInput());
         outState.putString("organiDesc", inputOrgiDesc.getInput());
-        outState.putSerializable("listData", (Serializable) departments);
+//        outState.putSerializable("listData", (Serializable) departments);
     }
 
     @Override
@@ -277,20 +295,21 @@ public class AddOrganizationFragment extends BaseFunctionFragment implements Ext
             orgaName = savedInstanceState.getString("organiName");
 
             orgaDesc = savedInstanceState.getString("organiDesc");
-            departments = (List<Organization>) savedInstanceState.get("listData");
-            if (newDepartment != null) {
-                departments.add(newDepartment);
-                newDepartment = null;
-            }
+//            departments = (List<Organization>) savedInstanceState.get("listData");
+//            if (newDepartment != null) {
+//                departments.add(newDepartment);
+//                newDepartment = null;
+//            }
         }
 
-        Logger.d("organiName = " + orgaName + "organiDesc = " + orgaDesc + "  list size =  " + departments.size());
+//        Logger.d("organiName = " + orgaName + "organiDesc = " + orgaDesc + "  list size =  " + departments.size());
 
-        exvAddOrgi.setListviewData(new ExtDeptAdapter(departments, true, (parentOrgisAdapter, position) -> {
-            deleteItem(parentOrgisAdapter, position);
-        }));
+        if (departments!=null){
+            exvAddOrgi.setListviewData(new ExtDeptAdapter(departments, true, (parentOrgisAdapter, position) -> {
+                deleteItem(parentOrgisAdapter, position);
+            }));
+        }
 
-//        }
     }
 
     @Override
